Change to version 2.00001
Pretty much a complete re-write of the whole thing. I will not bother
explaining it all here, but the broad strokes are improve everything by
removing stuff that I thought was good at the time. Now it's built with
browserify which is a pain, but I think makes good results. It seems
that any userscript that gets very complex will be a pain in my ass
eventually. Hopefully this helps. I'm rambling, it's better now I hope.